(#63) Atoms: move most to the Shared assembly
