+++ src/WpfMath.Shared/Atoms/CharAtom.cs
