+++ src/WpfMath.Shared/Atoms/CharSymbol.cs
