+++ src/WpfMath.Shared/Atoms/DummyAtom.cs
