+++ src/WpfMath.Shared/Atoms/FencedAtom.cs
