+++ src/WpfMath.Shared/Atoms/FixedCharAtom.cs
