+++ src/WpfMath.Shared/Atoms/FractionAtom.cs
