+++ src/WpfMath.Shared/Atoms/IRow.cs
