+++ src/WpfMath.Shared/Atoms/MatrixAtom.cs
