+++ src/WpfMath.Shared/Atoms/NullAtom.cs
