+++ src/WpfMath.Shared/Atoms/OverUnderDelimiter.cs
