+++ src/WpfMath.Shared/Atoms/OverlinedAtom.cs
