+++ src/WpfMath.Shared/Atoms/PhantomAtom.cs
