+++ src/WpfMath.Shared/Atoms/RowAtom.cs
