+++ src/WpfMath.Shared/Atoms/SpaceAtom.cs
