+++ src/WpfMath.Shared/Atoms/StyledAtom.cs
