+++ src/WpfMath.Shared/Atoms/SymbolAtom.cs
