+++ src/WpfMath.Shared/Atoms/TypedAtom.cs
