+++ src/WpfMath.Shared/Atoms/UnderlinedAtom.cs
