+++ src/WpfMath.Shared/Atoms/VerticalCenteredAtom.cs
