+++ src/WpfMath.Shared/TexSymbolParser.cs
