+++ src/WpfMath.Shared/WpfMath.Shared.csproj
@@ -19,8 +19,4 @@
         <InternalsVisibleTo Include="WpfMath.Tests" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="Atoms" />
-    </ItemGroup>
-
 </Project>
